+++ tests/test_recursion.py
@@ -329,7 +329,7 @@ def test_recursion_with_bad_item():
                 "https://schemas.stacspec.org/v1.0.0/item-spec/json-schema/item.json"
             ],
             "valid_stac": False,
-            "error_type": "ValidationError",
+            "error_type": "JSONSchemaValidationError",
             "error_message": "'id' is a required property of the root of the STAC object",
         },
     ]
@@ -351,7 +351,7 @@ def test_recursion_with_missing_collection_link():
             ],
             "valid_stac": False,
             "validation_method": "recursive",
-            "error_type": "ValidationError",
+            "error_type": "JSONSchemaValidationError",
             "error_message": "'simple-collection' should not be valid under {}. Error is in collection",
         },
     ]